Separate file and message methods from TelegramClient
